[core] Moved loggers into separate files loggers.h and loggers.cpp
