+++ srtcore/filelist.maf
@@ -12,6 +12,7 @@ epoll.cpp
 fec.cpp
 handshake.cpp
 list.cpp
+loggers.cpp
 md5.cpp
 packet.cpp
 packetfilter.cpp
@@ -52,6 +53,7 @@ epoll.h
 handshake.h
 list.h
 logging.h
+loggers.h
 md5.h
 netinet_any.h
 packet.h
