+++ srtcore/loggers.h
@@ -1,0 +1,15 @@
+#include "logging.h"
+
+namespace srt_logging
+{
+
+extern Logger
+    glog,
+//    blog,
+    mglog,
+    dlog,
+    tslog,
+    rxlog,
+    cclog;
+
+}
